+++ pczt/src/lib.rs
@@ -43,6 +43,8 @@ mod orchard;
 mod sapling;
 mod transparent;
 
+#[cfg(feature = "zcp-builder")]
+const SAPLING_TX_VERSION: u32 = 4;
 const V5_TX_VERSION: u32 = 5;
 const V5_VERSION_GROUP_ID: u32 = 0x26A7270A;